Qt: Move logging setup into core
Sharable between frontends this way.
