+++ pcsx2-qt/QtHost.h
@@ -41,10 +41,6 @@ namespace QtHost
 	void Shutdown();
 
 	void UpdateFolders();
-	void UpdateLogging();
-
-	/// Initializes early console logging (for printing command line arguments).
-	void InitializeEarlyConsole();
 
 	/// Sets batch mode (exit after game shutdown).
 	bool InBatchMode();
