+++ pcsx2/CMakeLists.txt
@@ -1069,6 +1069,7 @@ if(PCSX2_CORE)
 		Frontend/InputManager.cpp
 		Frontend/InputSource.cpp
 		Frontend/LayeredSettingsInterface.cpp
+		Frontend/LogSink.cpp
 		GSDumpReplayer.cpp
 		HostSettings.cpp
 		VMManager.cpp
@@ -1079,6 +1080,7 @@ if(PCSX2_CORE)
 		Frontend/InputManager.h
 		Frontend/InputSource.h
 		Frontend/LayeredSettingsInterface.h
+		Frontend/LogSink.h
 		GSDumpReplayer.h
 		HostSettings.h
 		VMManager.h)
